thunken för att hämta requests i househldslice och state requests sätts i case
