+++ store/household/householdSlice.ts
@@ -5,6 +5,7 @@ import {
   editHouseholdToDB,
   getHouseholdsFromDB,
 } from "../../api/household";
+import { getRequestByHouseholdIdFromDb } from "../../api/profile";
 import { Household, HouseholdRequest } from "../../types";
 
 export interface HouseholdState {
@@ -86,6 +87,27 @@ export const addHouseholdAsync = createAsyncThunk<
     return thunkAPI.rejectWithValue(error.message);
   }
 });
+
+export const getRequestByHouseholdIdsAsync = createAsyncThunk(
+  "profiles/getRequestByHouseholdIds",
+  async (householdIds: string[], thunkAPI) => {
+    try {
+      const fetchedRequests: HouseholdRequest[] = [];
+
+      await Promise.all(
+        householdIds.map(async (householdId) => {
+          const requests = await getRequestByHouseholdIdFromDb(householdId);
+          if (requests) {
+            fetchedRequests.push(...requests);
+          }
+        }),
+      );
+      return fetchedRequests;
+    } catch (error: any) {
+      return thunkAPI.rejectWithValue(error.message);
+    }
+  },
+);
 
 const householdSlice = createSlice({
   name: "household",
@@ -170,6 +192,14 @@ const householdSlice = createSlice({
       })
       .addCase(addHouseholdAsync.rejected, (state, action) => {
         console.log("error vid add household: ", action.payload);
+      })
+      .addCase(getRequestByHouseholdIdsAsync.fulfilled, (state, action) => {
+        if (action.payload) {
+          state.requests = action.payload;
+        }
+      })
+      .addCase(getRequestByHouseholdIdsAsync.rejected, (state, action) => {
+        console.log("error vid get requests: ", action.payload);
       });
   },
 });
